backend/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.104.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
h11==0.14.0
httptools==0.6.1
idna==3.6
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.2
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
protobuf==4.21.12
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
zipp==3.17.0
